Fix warnings in cpp files and some header files

+++ src/port/ShipUtils.cpp
@@ -39,7 +39,7 @@ uint64_t GetUnixTimestamp() {
     auto time = std::chrono::system_clock::now();
     auto since_epoch = time.time_since_epoch();
     auto millis = std::chrono::duration_cast<std::chrono::milliseconds>(since_epoch);
-    long now = millis.count();
+    long now = static_cast<long>(millis.count());
     return now;
 }
