+++ src/port/ui/UIWidgets.cpp
@@ -29,7 +29,7 @@ std::string WrappedText(const char* text, unsigned int charactersPerLine) {
             lastSpace = currentCharacter;
         }
 
-        if ((currentLineLength >= charactersPerLine) && (lastSpace >= 0)) {
+        if ((currentLineLength >= (int)charactersPerLine) && (lastSpace >= 0)) {
             newText[lastSpace] = '\n';
             currentLineLength = currentCharacter - lastSpace - 1;
             lastSpace = -1;
@@ -926,9 +926,9 @@ bool CVarColorPicker(const char* label, const char* cvarName, Color_RGBA8 defaul
                                   .Color(themeColor)
                                   .Size(UIWidgets::Sizes::Inline))) {
             colorVec = GetRandomValue();
-            color.r = fmin(fmax(colorVec.x * 255, 0), 255);
-            color.g = fmin(fmax(colorVec.y * 255, 0), 255);
-            color.b = fmin(fmax(colorVec.z * 255, 0), 255);
+            color.r = static_cast<uint8_t>(fmin(fmax(colorVec.x * 255, 0), 255));
+            color.g = static_cast<uint8_t>(fmin(fmax(colorVec.y * 255, 0), 255));
+            color.b = static_cast<uint8_t>(fmin(fmax(colorVec.z * 255, 0), 255));
             CVarSetColor(valueCVar.c_str(), color);
             CVarSetInteger(rainbowCVar.c_str(), 0); // On click disable rainbow mode.
             ShipInit::Init(rainbowCVar.c_str());
@@ -1171,7 +1171,7 @@ ImVec4 GetRandomValue() {
 }
 
 Color_RGBA8 RGBA8FromVec(ImVec4 vec) {
-    Color_RGBA8 color = { vec.x * 255, vec.y * 255, vec.z * 255, vec.w * 255 };
+    Color_RGBA8 color = { static_cast<uint8_t>(vec.x * 255), static_cast<uint8_t>(vec.y * 255), static_cast<uint8_t>(vec.z * 255), static_cast<uint8_t>(vec.w * 255) };
     return color;
 }
